Refractor location of EnchantmentGetters
